% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompassTemplate-package.R
\docType{package}
\name{CompassTemplate-package}
\alias{CompassTemplate}
\alias{CompassTemplate-package}
\title{CompassTemplate: Compass RMarkdown Template}
\description{
Creates a Compass Branded RMarkdown document.
}
\author{
\strong{Maintainer}: Shauna Heron \email{sheron@compassne.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
